requirements.txt
configparser==3.8.1
Pillow==6.1.0
pytesseract==0.3.0
